fix: make examples compatible with refactored movesupersindomain struct
