+++ examples/eurec4a1d/src/main_eurec4a1d.cpp
@@ -30,11 +30,12 @@
 #include <string_view>
 
 #include "zarr/dataset.hpp"
-#include "cartesiandomain/add_supers_at_domain_top.hpp"
+#include "cartesiandomain/movement/add_supers_at_domain_top.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
-#include "cartesiandomain/cartesianmotion.hpp"
+#include "cartesiandomain/movement/cartesian_motion.hpp"
+#include "cartesiandomain/movement/cartesian_movement.hpp"
 #include "cartesiandomain/createcartesianmaps.hpp"
-#include "cartesiandomain/null_boundary_conditions.hpp"
+#include "cartesiandomain/movement/null_boundary_conditions.hpp"
 #include "coupldyn_fromfile/fromfile_cartesian_dynamics.hpp"
 #include "coupldyn_fromfile/fromfilecomms.hpp"
 #include "gridboxes/gridboxmaps.hpp"
@@ -100,7 +101,7 @@ inline auto create_movement(const Config &config, const Timesteps &tsteps,
   // const auto boundary_conditions = NullBoundaryConditions{};
   const auto boundary_conditions = AddSupersAtDomainTop(config.get_addsupersatdomaintop());
 
-  return MoveSupersInDomain(gbxmaps, motion, boundary_conditions);
+  return cartesian_movement(gbxmaps, motion, boundary_conditions);
 }
 
 inline MicrophysicalProcess auto create_microphysics(const Config &config,
