+++ roughpaper/scratch/main.cpp
@@ -26,9 +26,10 @@
 
 #include "zarr/dataset.hpp"
 #include "./cleotypes_sizes.hpp"
+#include "cartesiandomain/movement/null_boundary_conditions.hpp"
+#include "cartesiandomain/movement/cartesian_movement.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "cartesiandomain/createcartesianmaps.hpp"
-#include "cartesiandomain/null_boundary_conditions.hpp"
 #include "coupldyn_fromfile/fromfile_cartesian_dynamics.hpp"
 #include "coupldyn_fromfile/fromfilecomms.hpp"
 #include "gridboxes/gridboxmaps.hpp"
@@ -137,7 +138,7 @@ inline auto create_movement(const CartesianMaps &gbxmaps) {
   const Motion<CartesianMaps> auto motion = NullMotion{};
   const auto boundary_conditions = NullBoundaryConditions{};
 
-  return MoveSupersInDomain(gbxmaps, motion, boundary_conditions);
+  return cartesian_movement(gbxmaps, motion, boundary_conditions);
 }
 
 template <GridboxMaps GbxMaps>
