+++ roughpaper/src/main_impl.hpp
@@ -32,15 +32,14 @@
 
 #include "zarr/dataset.hpp"
 #include "cartesiandomain/movement/add_supers_at_domain_top.hpp"
-#include "cartesiandomain/movement/cartesian_movement.hpp"
 #include "cartesiandomain/movement/cartesian_motion.hpp"
+#include "cartesiandomain/movement/cartesian_movement.hpp"
 #include "cartesiandomain/movement/null_boundary_conditions.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "cartesiandomain/createcartesianmaps.hpp"
 #include "coupldyn_fromfile/fromfile_cartesian_dynamics.hpp"
 #include "coupldyn_fromfile/fromfilecomms.hpp"
 #include "gridboxes/gridboxmaps.hpp"
-#include "gridboxes/movesupersindomain.hpp"
 #include "initialise/config.hpp"
 #include "initialise/init_all_supers_from_binary.hpp"
 #include "initialise/init_supers_from_binary.hpp"
@@ -123,7 +122,7 @@ template <GridboxMaps GbxMaps>
 inline auto create_movement(const Config &config, const Timesteps &tsteps, const GbxMaps &gbxmaps) {
   const auto motion = create_motion(tsteps.get_motionstep());
   const auto boundary_conditions = create_boundary_conditions(config);
-  const auto movement = cartestian_movement(gbxmaps, motion, boundary_conditions);
+  const auto movement = cartesian_movement(gbxmaps, motion, boundary_conditions);
   return movement;
 }
